# [BEAM-8151, BEAM-7848] Swap to using an unbounded thread pool which shrinks when threads are not being used.