@lukecwik 2019-09-04
R: @pabloem @TheNeuralBit 
CC: @aaltay @robertwb @Hannah-Jiang