@lukecwik 2019-09-04
I removed it from everywhere except for the test that you mentioned and the one place where we have logic. Note that the default is 5 * cpu count (and not unlimited).