@lukecwik 2019-09-05
Run Python PreCommit