@lukecwik 2019-09-06
Run Python PreCommit