@aaltay 2019-09-07
Precommit is failing with:

16:09:15 collapsing-thread-pool-executor 2018.6 has requirement futures==3.2.0, but you have futures 3.3.0.
16:09:15 ERROR: InvocationError for command /home/jenkins/jenkins-slave/workspace/beam_PreCommit_Python_Phrase/src/sdks/python/test-suites/tox/py2/build/srcs/sdks/python/target/.tox-py27-lint/py27-lint/bin/pip check (exited with code 1)
16:09:15 py27-lint run-test-post: commands[0] | /home/jenkins/jenkins-slave/workspace/beam_PreCommit_Python_Phrase/src/sdks/python/test-suites/tox/py2/build/srcs/sdks/python/scripts/run_tox_cleanup.sh
16:09:15 ___________________________________ summary ____________________________________
16:09:15 ERROR:   py27-lint: commands failed

The reason is a dependency conflict.
collapsing-thread-pool-executor requires futures == 3.2.0 for python 2.7 (https://github.com/ftpsolutions/collapsing-thread-pool-executor/blob/master/setup.py#L22)

Beam's setup.py accepts "'futures>=3.2.0,<4.0.0; python_version < "3.0"',"

We can do a few things:
- Update collapsing-thread-pool-executor to relax the upper bound for futures dependency
- Update beam to restict its futures dependency to use 3.2.0