@aaltay 2019-09-07
Proposed this change (https://github.com/ftpsolutions/collapsing-thread-pool-executor/pull/2) will if that will be accepted.