@TheNeuralBit 2019-09-09
Is https://github.com/ftpsolutions/collapsing-thread-pool-executor/issues/1 a concern?