@lukecwik 2019-09-13
Ahmet's change was merged, waiting on release to PyPi