@lukecwik 2019-10-07
Done, waiting on tests to pass and will merge.