@lukecwik 2019-10-10
The issue is that the collapsing-thread-pool-executor has a bug where during shutdown it tries to join worker threads but the worker threads have a default 30 second timeout waiting for work.