@lukecwik 2019-10-11
Started discussion looking for alternatives:
https://lists.apache.org/thread.html/03710e3d37cdbb7cd4711f9ffebf9741f5e88b6d595542246b2d14f2@%3Cdev.beam.apache.org%3E