@robertwb 2019-11-08
Could we perhaps get this in guarded by a flag so we can do performance testing while we wait for fixes and another release?