@lukecwik 2019-11-08
I think the complexity of guarding this with a flag will be much larger then rebasing when the library is updated once https://github.com/ftpsolutions/collapsing-thread-pool-executor/pull/4 is merged and released so I'm not seeing the value.