@lukecwik 2019-11-12
Run Python PreCommit