@lukecwik 2019-11-13
R: @robertwb @aaltay @angoenka 
This has now changed enough that it should be re-reviewed.