@aaltay 2019-11-14
Run Python PreCommit